requirements.txt
aiosqlite>=0.19.0
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
spotipy>=2.23.0
lyricsgenius>=3.0.0
psutil>=5.9.0
PyNaCl>=1.5.0
quart>=0.19.0
hypercorn>=0.16.0
cryptography>=41.0.0
aioquic>=1.0.0
davey>=0.1.0
bcrypt>=4.0.0
